backend/requirements.txt
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.3.0
click==8.1.8
colorama==0.4.6
ecdsa==0.19.1
exceptiongroup==1.2.2
fastapi==0.115.12
greenlet==3.1.1
h11==0.14.0
idna==3.10
Mako==1.3.10
MarkupSafe==3.0.2
passlib==1.7.4
pyasn1==0.4.8
pydantic==2.11.3
pydantic_core==2.33.1
python-jose==3.4.0
python-multipart==0.0.20
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.1
